backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
defusedxml==0.7.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
gunicorn==23.0.0
httpie==3.2.4
httplib2==0.22.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.6.3
packaging==25.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
PyJWT==2.9.0
pyparsing==3.2.3
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
setuptools==80.9.0
soupsieve==2.7
sqlparse==0.5.3
twilio==9.7.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1
